requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cmake==3.21.1.post1
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gTTS==2.2.3
idna==3.2
lxml==4.6.3
MouseInfo==0.1.3
multitasking==0.0.9
mysql-connector-python==8.0.26
numpy==1.21.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
pandas==1.3.1
pickle-mixin==1.0.2
Pillow==8.3.1
playsound==1.3.0
protobuf==3.17.3
PyAudio==0.2.11
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyjokes==0.6.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pyTelegramBotAPI==4.0.0
python-dateutil==2.8.2
python3-xlib==0.15
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
pywhatkit==5.1
randfacts==0.8.0
requests==2.26.0
secure-smtplib==0.1.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
urllib3==1.26.6
weathercom==1.0.5
wikipedia==1.4.0
yfinance==0.1.63
